Add cli API emulator
In this major commit I've introduced a emulator to the GPIE source. To build
the emulator, use make emucli.

The emulator requires you to have the emulation version of the GPIE library
installed, as the paths have to match. The emulator emulates a gpio sysfs
set-up under /tmp/gpie.

A few code corrections have been done too.

Please note that the code is quite messy in it's current state. Ex some var
names are different from one file to another whil their role is still the
same. Some functions have comments, others don't. You'll also see plenty of
TODOS, all of that is planned to be cleaned in upcoming commits.

As for now the emulation is running on a very primitive CLI that displays
which pins are active and what state and mode they're. The CLI version is
planned to allow attachable devices such as LEDs etc, in a module styled way.
A GUI version is planned once we got a half decent CLI version.

Signed-off-by: CTXz <ctx.xda@gmail.com>